refactor: update import paths for photo state management atoms
